Allow embedding configuration in the remote URL.
I'd like to be able to use recursive_remote to store repositories that
are used to provide crates via Cargo. Although it is not by default,
Cargo can be configured to use the command-line git, so remote helpers
will work.

Currently, the only source of configuration for recursive-remote is the
git-config itself, so it's not possible to set that through Cargo as far
as I can tell (I didn't look into whether environment variables would
work).

This commit adds functionality to embed the configuration keys used by
recursive-remote in the upstream URL. When recursive-remote is reading
its configuration, it will include the embedded configuration at the
`APP` priority, which will override the repo config.

These keys are only visible to recursive-remote, and only those keys it
uses are included. This also adds command line options to encode/decode
the config files, and an integration test of course.

This design allows (but does not require) embedding the encryption key
material directly into the repository URL, which has the obvious
security implications. As with the repository configuration, the key can
be set to a filesystem path instead, in which case the key will be read
from there, thus avoiding this concern.

An odd edge case is that when an empty encryption key is embedded,
recursive-remote will generate a new key, and write it to the repository
config. Specifying a non-existent file will generate a new key and store
it there, which makes more sense.