Fix tests.
Tests are currently failing non-deterministically. When there is only
one test running, it typically passes. I observe failures with as few as
two, but the more the more frequent.

Debugging a bit, it appears that when git shells out to
recursive_remote, it is sometimes closing the pipe before RR prints the
final empty line after pushing. Deleting that empty line seems to work.
Not sure if Git changed behavior or what, but this seems to fix it.